+++ projects/gate/src/lib/index.ts
@@ -1,1 +1,3 @@
 export { Core } from "./tree/core";
+export type { Node } from "./tree/node";
+export type { Edge } from "./tree/edge";
